- Move unit tests to new "GMLOffRoadWriterTest", since "GMLWriterTest" was a more generic test file
